requirements.txt
streamlit==1.49.0
python-dotenv==1.1.1
pandas==2.1.1
numpy==1.26.4
plotly==5.21.0
folium==0.16.0
streamlit-folium==0.11.0
streamlit-option-menu==0.4.0
h3==4.3.1
streamlit-authenticator==0.4.1
matplotlib==3.7.1
PyYAML
